// alpha-version/tasks-ui/src/descriptor-task/backend-types.ts
export type TaskId = string;


export interface Task {
  readonly created: string;
  readonly updated: string;
  readonly treeVersion: string;
  readonly archived: string | undefined;
  readonly startDate: string | undefined;
  readonly dueDate: string | undefined;

  readonly parentId: string | undefined;
  readonly transactions: TaskTransaction[];
  readonly roles: string[];
  readonly assigneeIds: string[];
  readonly reporterId: string;

  readonly title: string;
  readonly description: string;
  readonly priority: TaskPriority;
  readonly status: TaskStatus;
  readonly labels: string[];
  readonly extensions: TaskExtension[];
  readonly comments: TaskComment[];

  readonly checklist: Checklist[];

  readonly id: TaskId;
  readonly version: string;
  readonly documentType: 'TASK';
}


export type TaskStatus = 'CREATED' | 'IN_PROGRESS' | 'COMPLETED' | 'REJECTED';
export type TaskPriority = 'LOW' | 'MEDIUM' | 'HIGH';

export interface Checklist {
  id: string;
  title: string;
  items: ChecklistItem[];
}

export interface ChecklistItem {
  id: string;
  title: string;
  completed: boolean;
  assigneeIds: string[];
  dueDate: string | undefined;
}

export interface TaskTransaction {
  id: string;
  commands: TaskCommand[];
}

export interface TaskExtension {
  id: string;
  type: 'dialob' | 'upload' | 'CUSTOMER' | string ;
  externalId: string;
  body?: {};
  created: string;
  updated: string;
}

export interface TaskComment {
  id: string;
  created: Date;
  replyToId: string | undefined;
  commentText: string;
  username: string;
}

export interface TaskHistory {
  id: string;
  versions: Task[];
}

export interface TaskCommand {
  targetDate?: string;
  userId?: string;
  commandType: TaskCommandType;
}